update deno to 2.2.1

+++ denograd/env/deno.ts
@@ -17,5 +17,5 @@ export class DenoEnv extends Environment {
   execSync = (command: string, { args }: { args?: string[] } = {}) => new Deno.Command(command, { args }).outputSync()
   tmpdir = () => os.tmpdir()
   homedir = () => os.homedir()
-  sha256 = (data: string | Uint8Array) => createHash('sha256').update(data).digest()
+  sha256 = (data: string | Uint8Array) => createHash('sha256').update(data).digest() as Uint8Array
 }
